+++ packages/aws-amplify-angular/src/components/authenticator/forgot-password-component/forgot-password.component.core.ts
@@ -45,6 +45,7 @@ const template = `
         (keyup)="setCode(code.value)"
         class="amplify-form-input"
         type="text"
+        autocomplete="off"
         placeholder="{{ this.amplifyService.i18n().get('Enter code') }}"
         data-test="${auth.forgotPassword.codeInput}"
       />
@@ -59,6 +60,7 @@ const template = `
         (keyup.enter)="onSubmit()"
         class="amplify-form-input"
         type="password"
+        autocomplete="off"
         placeholder="{{ this.amplifyService.i18n().get('Password') }}"
         data-test="${auth.forgotPassword.newPasswordInput}"
       />