+++ packages/aws-amplify-react/src/Widget/Picker.tsx
@@ -66,10 +66,15 @@ export default class Picker extends Component<IPickerProps, {}> {
 		const accept = this.props.accept || '*/*';
 
 		const theme = this.props.theme || AmplifyTheme;
+		const pickerStyle = Object.assign(
+			{},
+			{ position: 'relative' },
+			theme.pickerPicker
+		);
 		const inputStyle = Object.assign({}, PickerInput, theme.pickerInput);
 
 		return (
-			<div style={theme.pickerPicker}>
+			<div style={pickerStyle}>
 				<PhotoPickerButton theme={theme}>{I18n.get(title)}</PhotoPickerButton>
 				<input
 					title={I18n.get(title)}