+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-react-native@2.2.2...aws-amplify-react-native@2.2.3) (2019-10-23)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
+## [2.2.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.2.0...aws-amplify-react-native@2.2.2) (2019-10-10)
+
+**Note:** Version bump only for package aws-amplify-react-native
+
+
+
+
+
+# [2.2.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.19...aws-amplify-react-native@2.2.0) (2019-10-10)
+
+
+### Features
+
+* Added Prettier formatting ([4dfd9aa](https://github.com/aws/aws-amplify/commit/4dfd9aa9ab900307c9d17c68448a6ca4aa08fd5a))
+
+
+
+
+
 ## [2.1.19](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.1.17...aws-amplify-react-native@2.1.19) (2019-09-05)
 
 **Note:** Version bump only for package aws-amplify-react-native