+++ packages/aws-amplify-vue/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.3.2...aws-amplify-vue@0.3.3) (2019-10-23)
+
+**Note:** Version bump only for package aws-amplify-vue
+
+
+
+
+
+## [0.3.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.3.0...aws-amplify-vue@0.3.2) (2019-10-10)
+
+**Note:** Version bump only for package aws-amplify-vue
+
+
+
+
+
+# [0.3.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.2.17...aws-amplify-vue@0.3.0) (2019-10-10)
+
+
+### Features
+
+* Added Prettier formatting ([4dfd9aa](https://github.com/aws-amplify/amplify-js/commit/4dfd9aa9ab900307c9d17c68448a6ca4aa08fd5a))
+
+
+
+
+
 ## [0.2.17](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.2.14...aws-amplify-vue@0.2.17) (2019-09-05)
 
 **Note:** Version bump only for package aws-amplify-vue